/@/store/domain/viewStatesの分離
